requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==3.14.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.7.0
idna==3.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
